:recycle: (Achievement) Reorganize folders and remove useless web tests
