+++ src/ui/components/achievements/index.ts
@@ -1,5 +1,5 @@
-import { AchievementCardKeyProps as AchievementCardKeyPropsType } from './components/GenericAchievementCard'
-export { GenericAchievement } from './components/GenericAchievement'
-export { GenericAchievementCard } from './components/GenericAchievementCard'
+import { AchievementCardKeyProps as AchievementCardKeyPropsType } from './components/GenericAchievementCard/GenericAchievementCard'
+export { GenericAchievement } from './components/GenericAchievement/GenericAchievement'
+export { GenericAchievementCard } from './components/GenericAchievementCard/GenericAchievementCard'
 
 export type AchievementCardKeyProps = AchievementCardKeyPropsType
